Changed OOCreditsQuantity to uint64_t.
git-svn-id: http://svn.berlios.de/svnroot/repos/oolite-linux/trunk@1274 127b21dd-08f5-0310-b4b7-95ae10353056

+++ src/Core/OOTypes.h
@@ -288,7 +288,7 @@ typedef int32_t		OOCargoQuantityDelta;
 typedef uint16_t	OOFuelQuantity;
 
 
-typedef uint32_t	OOCreditsQuantity;
+typedef uint64_t	OOCreditsQuantity;
 
 
 typedef uint16_t	OOKeyCode;
